lama/requirements/base.txt
Django==4.2.3
djangorestframework==3.14.0
python-decouple==3.8
psycopg2==2.9.6